MyFunctionapp/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-storage-blob
requests
python-dotenv